Split Balsam controller into its own module.
