+++ libensemble/tests/unit_tests/test_jobcontroller.py
@@ -7,7 +7,8 @@ import time
 import pytest
 import socket
 from libensemble.register import Register, BalsamRegister
-from libensemble.controller import JobController, BalsamJobController
+from libensemble.controller import JobController
+from libensemble.balsam_controller import BalsamJobController
 
 USE_BALSAM = False
